Refactor regclient config and interfaces
- Config file processing moved from regclient to regctl
- Refactoring interfaces for each client component to define its own interface
  and regclient is the assembly of all client components
