+++ cmd/regctl/config_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package main
+
+import "os"
+
+func getFileOwner(stat os.FileInfo) (int, int, error) {
+	return 0, 0, nil
+}
